requirements.txt
mysql-connector-python
pymongo
pandas